Refactor and sort code across the files.
